Refactor ExternalProcessScriptRuntime into multiple files for easier maintenance
